+++ node/src/routes/auth.module.css
@@ -1,3 +1,8 @@
+.title {
+    padding-bottom: 20px;
+    font-size: 2em;
+}
+
 .container {
 	display: flex;
 	flex-direction: column;
