+++ node/src/routes/auth/-components/RecipeCard.module.css
@@ -53,7 +53,7 @@
 .recipeImage {
     width: 100%;
     height: 20vh;
-    object-fit: cover;
+    object-fit: contain;
 }
 
 .recipeContent {
